[WCNCR00262566] wlan: refactor pointer casting
[Description]
refactor uintptr_t as unsigned long for different OS pointer type on x64/32
add define of uintptr_t if not in other OS

MTK-Commit-Id: da16bd5f8f9e49e388ed58b3805a97c3eeb4e398

CR-Id: WCNCR00262566
Feature: os/wlan
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6073805
Tested-by: yuchi.wang <yuchi.wang@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_mspautosanity
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>

+++ common/wlan_bow.c
@@ -522,11 +522,13 @@ uint32_t bowCmdSetupConnection(IN struct ADAPTER *prAdapter, IN struct BT_OVER_W
 
 		cnmTimerInitTimer(prAdapter,
 				  &prBowFsmInfo->rStartingBeaconTimer,
-				  (PFN_MGMT_TIMEOUT_FUNC) bowSendBeacon, (unsigned long) NULL);
+				  (PFN_MGMT_TIMEOUT_FUNC) bowSendBeacon,
+				  (uintptr_t) NULL);
 
 		cnmTimerInitTimer(prAdapter,
 				  &prBowFsmInfo->rChGrantedTimer,
-				  (PFN_MGMT_TIMEOUT_FUNC) bowChGrantedTimeout, (unsigned long) NULL);
+				  (PFN_MGMT_TIMEOUT_FUNC) bowChGrantedTimeout,
+				  (uintptr_t) NULL);
 
 		/* Reset Global Variable */
 		g_u4Beaconing = 0;
@@ -1722,7 +1724,8 @@ u_int8_t bowValidateProbeReq(IN struct ADAPTER *prAdapter, IN struct SW_RFB *prS
 	prMgtHdr = (struct WLAN_MAC_MGMT_HEADER *) prSwRfb->pvHeader;
 
 	u2IELength = prSwRfb->u2PacketLen - prSwRfb->u2HeaderLen;
-	pucIE = (uint8_t *) (((unsigned long) prSwRfb->pvHeader) + prSwRfb->u2HeaderLen);
+	pucIE = (uint8_t *) (((uintptr_t) prSwRfb->pvHeader) +
+		prSwRfb->u2HeaderLen);
 
 	IE_FOR_EACH(pucIE, u2IELength, u2Offset) {
 		if (IE_ID(pucIE) == ELEM_ID_SSID) {
@@ -1772,7 +1775,7 @@ u_int8_t bowValidateProbeReq(IN struct ADAPTER *prAdapter, IN struct SW_RFB *prS
 * @return (none)
 */
 /*----------------------------------------------------------------------------*/
-void bowSendBeacon(IN struct ADAPTER *prAdapter, IN unsigned long ulParamPtr)
+void bowSendBeacon(IN struct ADAPTER *prAdapter, IN uintptr_t ulParamPtr)
 {
 	struct BOW_FSM_INFO *prBowFsmInfo;
 
@@ -2913,7 +2916,7 @@ void bowReleaseCh(IN struct ADAPTER *prAdapter)
 * @return (none)
 */
 /*----------------------------------------------------------------------------*/
-void bowChGrantedTimeout(IN struct ADAPTER *prAdapter, IN unsigned long ulParamPtr)
+void bowChGrantedTimeout(IN struct ADAPTER *prAdapter, IN uintptr_t ulParamPtr)
 {
 	struct BOW_FSM_INFO *prBowFsmInfo;
 	enum ENUM_BOW_DEVICE_STATE eFsmState;
